Patch:  Rationalize RTAI configuration
The autoconf configuration script assumes that RTAI is built in the
source tree, and there exists an 'RTAI_DIR' containing
e.g. include/rtai_sched.h.

This patch replaces the '--with-rtai-dir' argument with
'--with-rtai-config', and instead of assuming the location of RTAI
headers relative to an 'RTAI_DIR', grabs the entire list of kbuild
EXTRA_CFLAGS from `rtai-config`.