Break out libethercat package and add Multi-Arch: tags
Break out a `libethercat` package so that other packages that build
against the `etherlabmaster-dev` package don't needlessly drag in the
DKMS config, which is problematic when building in a CI environment
where the running kernel doesn't match installed headers.

Add `Multi-Arch: foreign` tag to `etherlabmaster-dev` package so
`mk-build-deps -a $FOREIGN_ARCH` in the `linuxcnc-ethercat` package
pulls in the right `etherlabmaster-dev:$FOREIGN_ARCH` package.